requirements.txt
asgiref>=3.7.2
beautifulsoup4>=4.12.3
certifi>=2024.2.2
charset-normalizer>=3.3.2
click>=8.1.7
colorama>=0.4.6
fastapi==0.111.0
gunicorn>=21.2.0
h11>=0.14.0
httptools>=0.6.1
idna>=3.7
importlib-metadata>=7.1.0
lxml>=5.2.1
pydantic==2.5.2
python-dotenv==1.0.0
PyYAML>=6.0.1
requests>=2.32.2
soupsieve>=2.5
starlette==0.37.2
tqdm>=4.67.1
typing-extensions>=4.11.0
urllib3>=2.2.1
uvicorn==0.24.0
watchgod>=0.8.2
websockets>=12.0
zipp>=3.18.1
# Database dependencies
sqlalchemy>=2.0.0
psycopg2-binary==2.9.9
alembic>=1.13.0
# Optional dependencies for enhanced performance
redis==5.0.1  # For distributed caching (optional)
